backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
transformers
torch
soundfile
groq
python-multipart